chore: add GitHub community files and update README (#1)
* chore: add GitHub community files and update README

- Add SECURITY.md with vulnerability reporting policy
- Add dependabot.yml for automated dependency updates
- Update README with badges and quick start section
- Update license reference from MIT to Apache 2.0

* fix: fix dependabot ecosystems and update security contact email

* fix(lint): wrap email as mailto link to satisfy MD034

* fix(CI): use vendor binaries, cast OAuth TTLs to int, fix Pint issues, regenerate PHPStan baseline

* Chore(ignore) Exclude Claude.md

+++ config/core.php
@@ -98,7 +98,7 @@ return [
     | OAuth2 durees (en minutes)
     |--------------------------------------------------------------------------
     */
-    'oauth_access_token_ttl' => env('OAUTH_ACCESS_TOKEN_TTL', 60),
-    'oauth_refresh_token_ttl' => env('OAUTH_REFRESH_TOKEN_TTL', 43200),
+    'oauth_access_token_ttl' => (int) env('OAUTH_ACCESS_TOKEN_TTL', 60),
+    'oauth_refresh_token_ttl' => (int) env('OAUTH_REFRESH_TOKEN_TTL', 43200),
 
 ];
